requirements.txt
python-dotenv>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
httpx>=0.27.0
openpyxl>=3.1.0
pandas>=2.0.0
reportlab>=4.0.0
